refactor(dom): move helper to a file
